+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.4.9] - Unreleased
 
+### Added
+
+- Warning when a property is used and it is not in the binding
+- Support for `property-allowlist` for Zephyr bindings
+
 ### Changed
 
 - 'File not in context' downgraded from warning to information